Update README.md summery
The Anime Gear Destination website is designed as a hub for anime fans to discover and purchase anime-themed merchandise. The site is structured to cater to users who are passionate about anime and its culture, offering a wide range of products that bring the vibrant world of anime to everyday life. With a focus on providing an easy-to-navigate interface, attractive product displays, and detailed information about each item, this website aims to offer a seamless and engaging shopping experience for anime enthusiasts of all age